refactored user into user and login modules
