+++ src/store/modules/user/index.js
@@ -1,6 +1,5 @@
 import axios from 'axios'
-import util from 'util'
-import { SET_USERNAME, SET_TOKEN, SET_LOADING } from '@/store/mutation-types'
+// import { SET_USERNAME, SET_TOKEN } from '@/store/mutation-types'
 
 const server = 'http://cmpt371g4.usask.ca:8081'
 
@@ -15,51 +14,17 @@ const state = {
 const getters = {
   user (state) {
     return state.user
-  },
-  token (state) {
-    return state.token
-  },
-  isAuthenticated (state) {
-    return state.user !== null && state.user !== undefined
   }
 }
 
 // mutations of this module, mutation must be sync and atomic
 const mutations = {
-  [SET_USERNAME] (state, payload) {
-    state.user = payload
-  },
-  [SET_TOKEN] (state, payload) {
-    state.token = payload
-  }
+
 }
 
 // actions can be async and may have side effects
 const actions = {
   // payload is username and password
-  signIn ({ commit }, payload) {
-    commit(SET_LOADING, { loading: true })
-
-    return new Promise((resolve, reject) => {
-      axios.post(server + '/api/login', payload)
-        .then(function (responce) {
-          commit(SET_TOKEN, responce.token)
-          commit(SET_LOADING, { loading: false })
-          commit(SET_USERNAME, responce.username)
-        })
-        .catch(function (err) {
-          console.log(util.inspect(err, false, null))
-          reject(err)
-          commit(SET_LOADING, { loading: false })
-          console.log(err)
-        })
-    })
-  },
-  signOut ({ commit }) {
-    commit(SET_TOKEN, '')
-    alert('logged out')
-  },
-  // payload is usernname and password
   createUser (payload) {
     return new Promise((resolve, reject) => {
       axios.post(server + '/api/user/', payload)
